backend/requirements.txt
fastapi
uvicorn
langchain
openai
pydantic
aiofiles
python-dotenv
openai